laudatory-larkspurs/pyproject.toml
[project]
name = "good-image-terminal"
description = "A simple yet interesting image editor."
authors = [
    { name = "Mira"},
    { name = "Julien"},
    { name = "Jont"},
    { name = "Philip"},
    { name = "Richard Szilagyi", email = "szprichard@proton.me"},
    { name = "Mark"}
]
version = "0.1.0"
license = "MIT"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyodide-py>=0.27.7",
    "pillow~=11.3.0",
    "webcolors~=24.11.1"
]

[dependency-groups]
# This `dev` group contains all the development requirements for our linting toolchain.
# Don't forget to pin your dependencies!
# This list will have to be migrated if you wish to use another dependency manager.
dev = [
    "pyodide-py>=0.27.7",
    "pre-commit~=4.2.0",
    "ruff~=0.12.2",
    "pillow~=11.3.0",
    "webcolors~=24.11.1"
]

[tool.ruff]
# Increase the line length. This breaks PEP8 but it is way easier to work with.
# The original reason for this limit was a standard vim terminal is only 79 characters,
# but this doesn't really apply anymore.
line-length = 119
# Target Python 3.12. If you decide to use a different version of Python
# you will need to update this value.
target-version = "py312"
# Automatically fix auto-fixable issues.
fix = true
# The directory containing the source code. If you choose a different project layout
# you will need to update this value.
src = ["src"]

[tool.ruff.lint]
# Enable all linting rules.
select = ["ALL"]
# Ignore some of the most obnoxious linting errors.
ignore = [
    # Missing docstrings.
    "D100",
    "D104",
    "D105",
    "D106",
    "D107",
    # Docstring whitespace.
    "D203",
    "D213",
    # Docstring punctuation.
    "D415",
    # Docstring quotes.
    "D301",
    # Builtins.
    "A",
    # Print statements.
    "T20",
    # TODOs.
    "TD002",
    "TD003",
    "FIX",
]

[tool.ruff.lint.per-file-ignores]
# ignore Magic value and ambigus variable names when handling colors and color space conversions
"src/utils/color.py" = ["E741", "PLR2004"]
